unsigned long -> size_t in gauss

+++ linbox/algorithms/gauss/gauss-det.inl
@@ -34,11 +34,11 @@ namespace LinBox
 	template <class _Matrix> inline typename GaussDomain<_Field>::Element&
 	GaussDomain<_Field>::detin(Element        &determinant,
 				   _Matrix        &A,
-				   unsigned long  Ni,
-				   unsigned long  Nj,
+				   size_t  Ni,
+				   size_t  Nj,
 				   SparseEliminationTraits::PivotStrategy   reord)  const
 	{
-		unsigned long Rank;
+		size_t Rank;
 		if (reord == SparseEliminationTraits::PIVOT_NONE)
 			NoReordering(Rank, determinant, A,  Ni, Nj);
 		else
@@ -71,12 +71,12 @@ namespace LinBox
 	template <class _Matrix> inline typename GaussDomain<_Field>::Element&
 	GaussDomain<_Field>::det(Element       &determinant,
 				 const _Matrix  &A,
-				 unsigned long  Ni,
-				 unsigned long  Nj,
+				 size_t  Ni,
+				 size_t  Nj,
 				 SparseEliminationTraits::PivotStrategy   reord)  const
 	{
 		_Matrix CopyA(Ni);
-		for(unsigned long i = 0; i < Ni; ++i)
+		for(size_t i = 0; i < Ni; ++i)
 			CopyA[i] = A[i];
 		return detin(determinant, CopyA, Ni, Nj, reord);
 	}
